requirements.txt
asgiref==3.8.1
Brotli==1.1.0
Django==4.2.16
djangorestframework==3.15.2
fonttools==4.53.1
gunicorn==23.0.0
html5lib==1.1
packaging==24.1
pillow==10.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
webencodings==0.5.1
zopfli==0.2.3
